Conflicts:
	Code/Gui/MITK/Plugins/uk.ac.ucl.cmic.gui.qt.commonapps/CMakeLists.txt

+++ Code/Gui/MITK/Modules/IGI/Testing/mitkMergePointCloudsTest.cxx
@@ -98,7 +98,7 @@ static void TestSimpleCases()
 {
   mitk::PointSet::Pointer   one = mitk::PointSet::New();
   {
-    float   p[3] = {std::rand(), std::rand(), -2};
+    float   p[3] = {static_cast<float>(std::rand()), static_cast<float>(std::rand()), -2};
     one->InsertPoint(3, mitk::PointSet::PointType(&p[0]));
   }
 
@@ -115,7 +115,7 @@ static void TestSimpleCases()
   mitk::PointSet::Pointer   two = mitk::PointSet::New();
   for (int i = 0; i < 2; ++i)
   {
-    float   p[3] = {0, std::rand(), std::rand()};
+    float   p[3] = {0, static_cast<float>(std::rand()), static_cast<float>(std::rand())};
     two->InsertPoint(i, mitk::PointSet::PointType(&p[0]));
   }
